basholeveldb: Fix iterator example usage to not crash
The example used an "off by one" looping error. The iterator points to the first value after the seek, but the body called `Next()` before that first value was used. Also, because it's one too far forward, it would `Next()` past the last entry and calling `Value()` caused a panic.